Use separate headers instead of binary flags in the xhr sandbox (close #122)
