+++ src/const.js
@@ -18,10 +18,6 @@ CONST.HOVER_PSEUDO_CLASS_ATTR = 'data-hammerhead-hovered';
 
 CONST.DOCUMENT_CHARSET = 'hammerhead|document-charset';
 
-CONST.XHR_CORS_SUPPORTED_FLAG   = 0x10;
-CONST.XHR_REQUEST_MARKER_HEADER = 'hammerhead|xhr-request-marker-header';
-CONST.XHR_WITH_CREDENTIALS_FLAG = 0x01;
-
 CONST.SHADOW_UI_CLASSNAME_POSTFIX         = '-hammerhead-shadow-ui';
 CONST.SHADOW_UI_CHARSET_CLASSNAME         = 'charset' + CONST.SHADOW_UI_CLASSNAME_POSTFIX;
 CONST.SHADOW_UI_SCRIPT_CLASSNAME          = 'script' + CONST.SHADOW_UI_CLASSNAME_POSTFIX;
