+++ src/pipeline/xhr/headers.js
@@ -1,0 +1,10 @@
+// -------------------------------------------------------------
+// WARNING: this file is used by both the client and the server.
+// Do not use any browser or node-specific API!
+// -------------------------------------------------------------
+
+export default {
+    requestMarker:   'hammerhead|xhr|request-marker-header',
+    corsSupported:   'hammerhead|xhr|cors-supported-header',
+    withCredentials: 'hammerhead|xhr|with-credentials-header'
+};
